% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cf}
\alias{cf}
\title{Spearman calculation performed manually for a single sample}
\usage{
cf(counts)
}
\arguments{
\item{query.df}{A data frame containing one sample in a single column with ENSEMBL IDs as row names.  ENSEMBL IDs should not have decimals.}

\item{ref.df}{A data frame containing the averaged GTEx reference transcriptomes with ENSEMBL IDs.  IDs should not have decimals.}

\item{weighted}{If TRUE, delta squared values are weighted by rank.  Defaults to FALSE.}
}
\description{
Spearman calculation performed manually for a single sample
}
\examples{
spearman.manual.for.single.sample()
}
