% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot.spearman}
\alias{plot.spearman}
\title{Plot spearman results}
\usage{
\method{plot}{spearman}(spearman.results)
}
\arguments{
\item{query.dataframe}{A data frame containing samples in a each column with ENSEMBL IDs as row names.  ENSEMBL IDs should not have decimals.}

\item{reference.dataframe}{A data frame containing the averaged GTEx reference transcriptomes with ENSEMBL IDs.  IDs should not have decimals.}

\item{method}{Options are 'corr', 'manual', and 'weighted'.  Defaults to 'corr'.}

\item{order.by.score}{If TRUE, outputs a table reordered by average score.  If FALSE, output table is in same order as reference data frame columns.  Defaults to TRUE}
}
\description{
Plot spearman results
}
\examples{
spearman.calc()
}
