% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{spearman.calc}
\alias{spearman.calc}
\title{Spearman calculation for multiple samples}
\usage{
spearman.calc(query.df, ref.df, method = "corr", order.by.score = TRUE)
}
\arguments{
\item{method}{Options are 'corr', 'manual', and 'weighted'.  Defaults to 'corr'.}

\item{order.by.score}{If TRUE, outputs a table reordered by average score.  If FALSE, output table is in same order as reference data frame columns.  Defaults to TRUE}

\item{query.dataframe}{A data frame containing samples in a each column with ENSEMBL IDs as row names.  ENSEMBL IDs should not have decimals.}

\item{reference.dataframe}{A data frame containing the averaged GTEx reference transcriptomes with ENSEMBL IDs.  IDs should not have decimals.}
}
\description{
Spearman calculation for multiple samples
}
\examples{
spearman.calc()
}
