% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{spearman.corr.for.single.sample}
\alias{spearman.corr.for.single.sample}
\title{Spearman calculation using rcorr function for a single sample}
\usage{
spearman.corr.for.single.sample(query.df, ref.df)
}
\arguments{
\item{query.df}{A data frame containing one sample in a single column with ENSEMBL IDs as row names.  ENSEMBL IDs should not have decimals.}

\item{ref.df}{A data frame containing the averaged GTEx reference transcriptomes with ENSEMBL IDs.  IDs should not have decimals.}
}
\description{
Spearman calculation using rcorr function for a single sample
}
\examples{
spearman.corr.for.single.sample()
}
